fix :url property in elpa/nongnu package lists
Elpa has started to put symbols in the :url property of some package
specs in its package list. This symbol is supposed to refer to another
package (an entry) in same package list. The package referred to will
have a real URL (a string) in the :url property -- or so we hope.

There is a new utility function which will perform this lookup and
replace the symbol URL with its real URL.